refactor(bridge): split repair/tests.rs into focused submodules (no behavior change)
Final-review note on PR #390: repair/tests.rs had grown to 1,234 lines,
past the 500-line split threshold. Verbatim code motion into
repair/tests/{mod,adopt,rollback_gate,consent,preview}.rs — mod.rs keeps
the shared fixture helpers (golden-vault staging, crashed-save/share
stagers, external-card mint) and each submodule takes one concern:

- adopt.rs         happy adopts + default widening refusal + idempotence
- rollback_gate.rs §10 pre-write fail-closed postures (#384)
- consent.rs       approvals on the three repair arms (#374 part 3)
- preview.rs       read-only preview arms (#374 Task 6)

Every test moved unchanged (all 18 still run: 4+6+5+3); only module
scaffolding (submodule decls, //! headers, `use super::*;`) is new.
Verified: cargo test --release -p secretary-ffi-bridge 177/177, full
workspace green, clippy/fmt/rustdoc -D warnings clean, Swift + Kotlin
conformance 27/27 (UDL doc-comment change in the sibling commit).
